fix: ButtonをitemDetailCardの外に配置するようにする
